+++ dashboard/pages/GOG.py
@@ -1,204 +1,16 @@
 '''A script that creates the GOG page for the dashboard.'''
 
 from os import environ as ENV
-from datetime import datetime, timedelta
-
-import pandas as pd
-import altair as alt
 import streamlit as st
 from dotenv import load_dotenv
 
-from psycopg2 import connect
-from psycopg2.extras import RealDictCursor
-from psycopg2.extensions import connection
 
-
-def get_db_connection(config) -> connection:
-    """Returns a connection to the database."""
-
-    return connect(
-        dbname=config["DB_NAME"],
-        user=config["DB_USER"],
-        password=config["DB_PASSWORD"],
-        host=config["DB_HOST"],
-        port=config["DB_PORT"],
-        cursor_factory=RealDictCursor
-    )
-
-
-def get_week_list() -> tuple:
-    """Returns a tuple containing the dates 
-    of the last seven days (not including today)."""
-    w_list = []
-
-    for each in range(1, 8):
-        day = datetime.now() - timedelta(days=each)
-        w_list.append(day.strftime('%Y-%m-%d'))
-
-    return tuple(w_list)
-
-
-def filter_tags(data_df: pd.DataFrame, tags_: list, col: str) -> pd.DataFrame:
-    """Returns a Data-frame that only has the relevant tags."""
-
-    data_df = data_df[data_df[col].isin(tags_)]
-
-    return data_df
-
-
-def filter_dates(data_df: pd.DataFrame, dates: list, col: str) -> pd.DataFrame:
-    """Returns a Data-frame that only has data from the last 7 days."""
-
-    data_df[col] = data_df[col].astype(str)
-    data_df = data_df[data_df[col].isin(dates)]
-
-    return data_df
-
-
-def metric_games_yest(conn_: connection) -> pd.DataFrame:
-    """Returns a Data-frame of all the games from the yesterday."""
-
-    yesterday = (datetime.now() - timedelta(days=1)).strftime('%Y-%m-%d')
-
-    with conn_.cursor() as cur:
-        cur.execute(f""" SELECT name, rating, price, release_date
-                    FROM game
-                    WHERE website_id = 2 AND release_date = '{yesterday}';""")
-        gog_games = cur.fetchall()
-
-    return pd.DataFrame(gog_games)
-
-
-def metrics_for_graphs_price(conn_: connection) -> pd.DataFrame:
-    """Returns a Data-frame of all the average prices for the last week."""
-    w_list = get_week_list()
-
-    with conn_.cursor() as cur:
-        cur.execute(f""" SELECT AVG(price), release_date
-                    FROM game
-                    WHERE website_id = 2 and release_date in {w_list}
-                    GROUP BY release_date;""")
-        gog_games = cur.fetchall()
-
-    return pd.DataFrame(gog_games)
-
-
-def metrics_for_graphs_count(conn_: connection) -> pd.DataFrame:
-    """Returns a Data-frame of all the number of games for the last week."""
-    w_list = get_week_list()
-
-    with conn_.cursor() as cur:
-        cur.execute(f""" SELECT COUNT(name), release_date
-                    FROM game
-                    WHERE website_id = 2 and release_date in {w_list}
-                    GROUP BY release_date;""")
-        gog_games = cur.fetchall()
-
-    return pd.DataFrame(gog_games)
-
-
-def metrics_for_graphs_rating(conn_: connection) -> pd.DataFrame:
-    """Returns a Data-frame of all the game ratings for the past week."""
-    w_list = get_week_list()
-
-    with conn_.cursor() as cur:
-        cur.execute(f""" SELECT AVG(rating), release_date
-                    FROM game
-                    WHERE website_id = 2 and release_date in {w_list}
-                    GROUP BY release_date;""")
-        gog_games = cur.fetchall()
-
-    return pd.DataFrame(gog_games)
-
-
-def metrics_for_graphs_tags(conn_: connection) -> pd.DataFrame:
-    """Returns a Data-frame of all the games tag data from the past week."""
-    w_list = get_week_list()
-
-    with conn_.cursor() as cur:
-        cur.execute(f"""SELECT t.tag_name, count(*) from tag AS t 
-                    INNER JOIN game_tag_matching AS gt 
-                    ON t.tag_id = gt.tag_id 
-                    WHERE gt.game_id IN
-                    (SELECT game_id from game 
-                    WHERE website_id = 2 AND release_date 
-                    IN {w_list}) 
-                    GROUP BY t.tag_name 
-                    LIMIT 10;""")
-        tag = cur.fetchall()
-
-    return pd.DataFrame(tag)
-
-
-def metrics_top_ten(conn_: connection) -> pd.DataFrame:
-    """Returns a Data-frame of top rated the games from the last week. """
-    w_list = get_week_list()
-
-    with conn_.cursor() as cur:
-        cur.execute(f"""SELECT g.name, g.rating, g.price, d.developer_name, p.publisher_name
-                    FROM game as g
-                    JOIN developer as d
-                    ON g.developer_id = d.developer_id
-                    JOIN publisher as p
-                    on g.publisher_id = p.publisher_id
-                    WHERE g.website_id = 2 and g.release_date in {w_list}
-                    ORDER BY rating DESC LIMIT 10;""")
-        tags_ = cur.fetchall()
-
-    return pd.DataFrame(tags_)
-
-
-def price_chart(data_df: pd.DataFrame, sorted_=True) -> alt.Chart:
-    """"Generates a bar chart of average daily prices of games over their release dates."""
-
-    data_df['release_date'] = data_df['release_date'].astype(str)
-    data_df['AVG price (£)'] = data_df['avg'].apply(lambda x: round(x, 2))
-    sort = "-y" if sorted_ else "x"
-
-    return alt.Chart(data_df).mark_bar().encode(
-        x=alt.Y("AVG price (£)", title='Average Daily Game Price'),
-        y=alt.X("release_date").sort(sort),
-        color="release_date"
-    )
-
-
-def count_chart(data_df: pd.DataFrame, sorted_=True) -> alt.Chart:
-    """"Generates a bar chart of daily number of games releases."""
-
-    data_df['release_date'] = data_df['release_date'].astype(str)
-    data_df['Daily Releases'] = data_df['count']
-    sort = "-y" if sorted_ else "x"
-
-    return alt.Chart(data_df).mark_line().encode(
-        x=alt.X("release_date",  title='Number Of Daily Releases').sort(sort),
-        y=alt.Y("Daily Releases")
-    )
-
-
-def rating_chart(data_df: pd.DataFrame, sorted_=True) -> alt.Chart:
-    """"Generates a bar chart of average daily ratings of games over their release dates."""
-
-    data_df['release_date'] = data_df['release_date'].astype(str)
-    data_df['Average Rating(%)'] = data_df['avg']
-    sort = "-y" if sorted_ else "x"
-
-    return alt.Chart(data_df).mark_bar().encode(
-        x=alt.Y("Average Rating(%)", title='Average Daily Game Rating'),
-        y=alt.X("release_date").sort(sort),
-        color="release_date"
-    )
-
-
-def make_tag_chart(data_df: pd.DataFrame, sorted_=True) -> alt.Chart:
-    """Generates a bar chart of most popular tags of games in the last week."""
-
-    sort = "-y" if sorted_ else "x"
-
-    return alt.Chart(data_df).mark_bar().encode(
-        x=alt.X("tag_name").sort(sort),
-        y="count",
-        color="tag_name"
-    )
+from pages.functions import (get_db_connection, get_week_list,
+                             make_tag_chart, metric_games_yest,
+                             metrics_for_graphs_count, metrics_for_graphs_price,
+                             metrics_for_graphs_rating, metrics_for_graphs_tags,
+                             metrics_top_ten, rating_chart, count_chart, price_chart,
+                             filter_dates, filter_tags)
 
 
 if __name__ == "__main__":
@@ -207,9 +19,9 @@ if __name__ == "__main__":
     conn = get_db_connection(ENV)
     week_list = list(get_week_list())
 
-    metric_df = metric_games_yest(conn)
-    top_ten_games = metrics_top_ten(conn)
-    tag_df = metrics_for_graphs_tags(conn)
+    metric_df = metric_games_yest(conn, 2)
+    top_ten_games = metrics_top_ten(conn, 2)
+    tag_df = metrics_for_graphs_tags(conn, 2)
     tags = tag_df["tag_name"].to_list()
 
     if not metric_df.empty:
@@ -221,9 +33,9 @@ if __name__ == "__main__":
         avg_rating = 0
         avg_price = 0
 
-    price_df = metrics_for_graphs_price(conn)
-    count_df = metrics_for_graphs_count(conn)
-    rating_df = metrics_for_graphs_rating(conn)
+    price_df = metrics_for_graphs_price(conn, 2)
+    count_df = metrics_for_graphs_count(conn, 2)
+    rating_df = metrics_for_graphs_rating(conn, 2)
 
     conn.close()
 
@@ -300,4 +112,3 @@ if __name__ == "__main__":
 
     st.subheader("Daily Releases")
     st.altair_chart(c_chart, use_container_width=True)
-
